Refactor the current vSphere provider structure
