+++ pkg/provider/cloud/vsphere/network.go
@@ -18,8 +18,10 @@ package vsphere
 
 import (
 	"context"
+	"crypto/x509"
 	"errors"
 	"fmt"
+	kubermaticv1 "k8c.io/kubermatic/v2/pkg/apis/kubermatic/v1"
 	"path"
 	"strings"
 
@@ -33,6 +35,21 @@ type NetworkInfo struct {
 	RelativePath string
 	AbsolutePath string
 	Type         string
+}
+
+// GetNetworks returns a slice of VSphereNetworks of the datacenter from the passed cloudspec.
+func GetNetworks(ctx context.Context, dc *kubermaticv1.DatacenterSpecVSphere, username, password string, caBundle *x509.CertPool) ([]NetworkInfo, error) {
+	// For the GetNetworks request we use dc.Spec.VSphere.InfraManagementUser
+	// if set because that is the user which will ultimatively configure
+	// the networks - But it means users in the UI can see vsphere
+	// networks without entering credentials
+	session, err := newSession(ctx, dc, username, password, caBundle)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create vCenter session: %w", err)
+	}
+	defer session.Logout(ctx)
+
+	return getPossibleVMNetworks(ctx, session)
 }
 
 func getPossibleVMNetworks(ctx context.Context, session *Session) ([]NetworkInfo, error) {
