+++ pkg/provider/cloud/vsphere/provider.go
@@ -21,24 +21,11 @@ import (
 	"crypto/x509"
 	"errors"
 	"fmt"
-	"net/url"
-	"path"
-	"strings"
-
-	"github.com/vmware/govmomi"
-	"github.com/vmware/govmomi/find"
-	"github.com/vmware/govmomi/object"
-	"github.com/vmware/govmomi/session"
-	"github.com/vmware/govmomi/vapi/tags"
-	"github.com/vmware/govmomi/vim25"
-	"github.com/vmware/govmomi/vim25/soap"
-
+	"go.uber.org/zap"
 	kubermaticv1 "k8c.io/kubermatic/v2/pkg/apis/kubermatic/v1"
 	kuberneteshelper "k8c.io/kubermatic/v2/pkg/kubernetes"
 	"k8c.io/kubermatic/v2/pkg/provider"
-	"k8c.io/kubermatic/v2/pkg/resources"
-
-	kruntime "k8s.io/apimachinery/pkg/util/runtime"
+	"path"
 )
 
 const (
@@ -46,106 +33,37 @@ const (
 	// categoryCleanupFinilizer will instruct the deletion of the default category tag.
 	tagCategoryCleanupFinilizer = "kubermatic.k8c.io/cleanup-vsphere-tag-category"
 
-	defaultCategory = "cluster"
+	defaultCategoryPrefix = "kubermatic.k8c.io"
 )
 
-// Provider represents the vsphere provider.
-type Provider struct {
+// VSphere represents the vsphere provider.
+type VSphere struct {
 	dc                *kubermaticv1.DatacenterSpecVSphere
+	log               *zap.SugaredLogger
 	secretKeySelector provider.SecretKeySelectorValueFunc
 	caBundle          *x509.CertPool
 }
 
-// Folder represents a vsphere folder.
-type Folder struct {
-	Path string
-}
-
 // NewCloudProvider creates a new vSphere provider.
-func NewCloudProvider(dc *kubermaticv1.Datacenter, secretKeyGetter provider.SecretKeySelectorValueFunc, caBundle *x509.CertPool) (*Provider, error) {
+func NewCloudProvider(dc *kubermaticv1.Datacenter, secretKeyGetter provider.SecretKeySelectorValueFunc, caBundle *x509.CertPool) (*VSphere, error) {
 	if dc.Spec.VSphere == nil {
 		return nil, errors.New("datacenter is not a vSphere datacenter")
 	}
-	return &Provider{
+	return &VSphere{
 		dc:                dc.Spec.VSphere,
 		secretKeySelector: secretKeyGetter,
 		caBundle:          caBundle,
 	}, nil
 }
 
-var _ provider.CloudProvider = &Provider{}
+var _ provider.ReconcilingCloudProvider = &VSphere{}
 
-type Session struct {
-	Client     *govmomi.Client
-	Finder     *find.Finder
-	Datacenter *object.Datacenter
-}
-
-// Logout closes the idling vCenter connections.
-func (s *Session) Logout(ctx context.Context) {
-	if err := s.Client.Logout(ctx); err != nil {
-		kruntime.HandleError(fmt.Errorf("vSphere client failed to logout: %w", err))
-	}
-}
-
-func newSession(ctx context.Context, dc *kubermaticv1.DatacenterSpecVSphere, username, password string, caBundle *x509.CertPool) (*Session, error) {
-	u, err := url.Parse(fmt.Sprintf("%s/sdk", dc.Endpoint))
-	if err != nil {
-		return nil, err
-	}
-
-	// creating the govmoni Client in roundabout way because we need to set the proper CA bundle: reference https://github.com/vmware/govmomi/issues/1200
-	soapClient := soap.NewClient(u, dc.AllowInsecure)
-	// set our CA bundle
-	soapClient.DefaultTransport().TLSClientConfig.RootCAs = caBundle
-
-	vim25Client, err := vim25.NewClient(ctx, soapClient)
-	if err != nil {
-		return nil, err
-	}
-
-	client := &govmomi.Client{
-		Client:         vim25Client,
-		SessionManager: session.NewManager(vim25Client),
-	}
-
-	user := url.UserPassword(username, password)
-	if dc.InfraManagementUser != nil {
-		user = url.UserPassword(dc.InfraManagementUser.Username, dc.InfraManagementUser.Password)
-	}
-
-	if err = client.Login(ctx, user); err != nil {
-		return nil, fmt.Errorf("failed to login: %w", err)
-	}
-
-	finder := find.NewFinder(client.Client, true)
-	datacenter, err := finder.Datacenter(ctx, dc.Datacenter)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get vSphere datacenter %q: %w", dc.Datacenter, err)
-	}
-	finder.SetDatacenter(datacenter)
-
-	return &Session{
-		Datacenter: datacenter,
-		Finder:     finder,
-		Client:     client,
-	}, nil
-}
-
-// getVMRootPath is a helper func to get the root path for VM's
-// We extracted it because we use it in several places.
-func getVMRootPath(dc *kubermaticv1.DatacenterSpecVSphere) string {
-	// Each datacenter root directory for VM's is: ${DATACENTER_NAME}/vm
-	rootPath := path.Join("/", dc.Datacenter, "vm")
-	// We offer a different root path though in case people would like to store all Kubermatic VM's below a certain directory
-	if dc.RootPath != "" {
-		rootPath = path.Clean(dc.RootPath)
-	}
-	return rootPath
+func (a *VSphere) ReconcileCluster(ctx context.Context, cluster *kubermaticv1.Cluster, update provider.ClusterUpdater) (*kubermaticv1.Cluster, error) {
+	return nil, nil
 }
 
 // InitializeCloudProvider initializes the vsphere cloud provider by setting up vm folders for the cluster.
-func (v *Provider) InitializeCloudProvider(ctx context.Context, cluster *kubermaticv1.Cluster, update provider.ClusterUpdater) (*kubermaticv1.Cluster, error) {
+func (v *VSphere) InitializeCloudProvider(ctx context.Context, cluster *kubermaticv1.Cluster, update provider.ClusterUpdater) (*kubermaticv1.Cluster, error) {
 	username, password, err := GetCredentialsForCluster(cluster.Spec.Cloud, v.secretKeySelector, v.dc)
 	if err != nil {
 		return nil, err
@@ -172,84 +90,50 @@ func (v *Provider) InitializeCloudProvider(ctx context.Context, cluster *kuberma
 			return nil, err
 		}
 	}
-	if cluster.Spec.Cloud.VSphere.TagCategoryID == "" {
+
+	tagCategoryName := cluster.Spec.Cloud.VSphere.TagCategoryName
+	// We only need to fetch/create tag categories only if the user explicitly decides to use on.
+	if tagCategoryName != "" {
 		restSession, err := newRESTSession(ctx, v.dc, username, password, v.caBundle)
 		if err != nil {
 			return nil, fmt.Errorf("failed to create REST client session: %w", err)
 		}
 		defer restSession.Logout(ctx)
 
-		// If the user did not specify a tag category, we create an own default for this cluster
-		categoryID, err := createTagCategory(ctx, restSession, cluster)
+		categoryID, err := fetchTagCategory(ctx, restSession, tagCategoryName)
+		if err != nil {
+			return nil, fmt.Errorf("failed to fetch tag category: %w", err)
+		}
+
+		if categoryID != "" {
+			return update(ctx, cluster.Name, func(cluster *kubermaticv1.Cluster) {
+				cluster.Spec.Cloud.VSphere.TagCategoryID = categoryID
+			})
+		}
+
+		categoryID, err = createTagCategory(ctx, restSession, cluster)
 		if err != nil {
 			return nil, fmt.Errorf("failed to create tag category: %w", err)
 		}
-		cluster, err = update(ctx, cluster.Name, func(cluster *kubermaticv1.Cluster) {
+
+		return update(ctx, cluster.Name, func(cluster *kubermaticv1.Cluster) {
 			kuberneteshelper.AddFinalizer(cluster, tagCategoryCleanupFinilizer)
 			cluster.Spec.Cloud.VSphere.TagCategoryID = categoryID
 		})
-		if err != nil {
-			return nil, err
-		}
 	}
 
 	return cluster, nil
 }
 
 // TODO: Hey, you! Yes, you! Why don't you implement reconciling for vSphere? Would be really cool :)
-// func (v *Provider) ReconcileCluster(cluster *kubermaticv1.Cluster, update provider.ClusterUpdater) (*kubermaticv1.Cluster, error) {
+// func (v *VSphere) ReconcileCluster(cluster *kubermaticv1.Cluster, update provider.ClusterUpdater) (*kubermaticv1.Cluster, error) {
 // 	return cluster, nil
 // }
 
-// GetNetworks returns a slice of VSphereNetworks of the datacenter from the passed cloudspec.
-func GetNetworks(ctx context.Context, dc *kubermaticv1.DatacenterSpecVSphere, username, password string, caBundle *x509.CertPool) ([]NetworkInfo, error) {
-	// For the GetNetworks request we use dc.Spec.VSphere.InfraManagementUser
-	// if set because that is the user which will ultimatively configure
-	// the networks - But it means users in the UI can see vsphere
-	// networks without entering credentials
-	session, err := newSession(ctx, dc, username, password, caBundle)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create vCenter session: %w", err)
-	}
-	defer session.Logout(ctx)
-
-	return getPossibleVMNetworks(ctx, session)
-}
-
-// GetVMFolders returns a slice of VSphereFolders of the datacenter from the passed cloudspec.
-func GetVMFolders(ctx context.Context, dc *kubermaticv1.DatacenterSpecVSphere, username, password string, caBundle *x509.CertPool) ([]Folder, error) {
-	session, err := newSession(ctx, dc, username, password, caBundle)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create vCenter session: %w", err)
-	}
-	defer session.Logout(ctx)
-
-	// We simply list all folders & filter out afterwards.
-	// Filtering here is not possible as vCenter only lists the first level when giving a path.
-	// vCenter only lists folders recursively if you just specify "*".
-	folderRefs, err := session.Finder.FolderList(ctx, "*")
-	if err != nil {
-		return nil, fmt.Errorf("couldn't retrieve folder list: %w", err)
-	}
-
-	rootPath := getVMRootPath(dc)
-	var folders []Folder
-	for _, folderRef := range folderRefs {
-		// We filter by rootPath. If someone configures it, we should respect it.
-		if !strings.HasPrefix(folderRef.InventoryPath, rootPath+"/") && folderRef.InventoryPath != rootPath {
-			continue
-		}
-		folder := Folder{Path: folderRef.Common.InventoryPath}
-		folders = append(folders, folder)
-	}
-
-	return folders, nil
-}
-
 // DefaultCloudSpec adds defaults to the cloud spec.
-func (v *Provider) DefaultCloudSpec(_ context.Context, spec *kubermaticv1.ClusterSpec) error {
-	if spec.Cloud.VSphere.TagCategoryID == "" {
-		spec.Cloud.VSphere.TagCategoryID = v.dc.DefaultTagCategoryID
+func (v *VSphere) DefaultCloudSpec(_ context.Context, spec *kubermaticv1.CloudSpec) error {
+	if spec.VSphere.TagCategoryID == "" {
+		spec.VSphere.TagCategoryID = v.dc.DefaultTagCategoryID
 	}
 
 	return nil
@@ -257,7 +141,7 @@ func (v *Provider) DefaultCloudSpec(_ context.Context, spec *kubermaticv1.Cluste
 
 // ValidateCloudSpec validates whether a vsphere client can be constructed for
 // the passed cloudspec and perform some additional checks on datastore config.
-func (v *Provider) ValidateCloudSpec(ctx context.Context, spec kubermaticv1.CloudSpec) error {
+func (v *VSphere) ValidateCloudSpec(ctx context.Context, spec kubermaticv1.CloudSpec) error {
 	username, password, err := GetCredentialsForCluster(spec, v.secretKeySelector, v.dc)
 	if err != nil {
 		return err
@@ -320,7 +204,7 @@ func (v *Provider) ValidateCloudSpec(ctx context.Context, spec kubermaticv1.Clou
 // CleanUpCloudProvider we always check if the folder is there and remove it if yes because we know its absolute path
 // This covers cases where the finalizer was not added
 // We also remove the finalizer if either the folder is not present or we successfully deleted it.
-func (v *Provider) CleanUpCloudProvider(ctx context.Context, cluster *kubermaticv1.Cluster, update provider.ClusterUpdater) (*kubermaticv1.Cluster, error) {
+func (v *VSphere) CleanUpCloudProvider(ctx context.Context, cluster *kubermaticv1.Cluster, update provider.ClusterUpdater) (*kubermaticv1.Cluster, error) {
 	username, password, err := GetCredentialsForCluster(cluster.Spec.Cloud, v.secretKeySelector, v.dc)
 	if err != nil {
 		return nil, err
@@ -365,7 +249,7 @@ func (v *Provider) CleanUpCloudProvider(ctx context.Context, cluster *kubermatic
 }
 
 // ValidateCloudSpecUpdate verifies whether an update of cloud spec is valid and permitted.
-func (v *Provider) ValidateCloudSpecUpdate(_ context.Context, oldSpec kubermaticv1.CloudSpec, newSpec kubermaticv1.CloudSpec) error {
+func (v *VSphere) ValidateCloudSpecUpdate(_ context.Context, oldSpec kubermaticv1.CloudSpec, newSpec kubermaticv1.CloudSpec) error {
 	if oldSpec.VSphere == nil || newSpec.VSphere == nil {
 		return errors.New("'vsphere' spec is empty")
 	}
@@ -375,118 +259,4 @@ func (v *Provider) ValidateCloudSpecUpdate(_ context.Context, oldSpec kubermatic
 	}
 
 	return nil
-}
-
-// GetDatastoreList returns a slice of Datastore of the datacenter from the passed cloudspec.
-func GetDatastoreList(ctx context.Context, dc *kubermaticv1.DatacenterSpecVSphere, username, password string, caBundle *x509.CertPool) ([]*object.Datastore, error) {
-	session, err := newSession(ctx, dc, username, password, caBundle)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create vCenter session: %w", err)
-	}
-	defer session.Logout(ctx)
-
-	datastoreList, err := session.Finder.DatastoreList(ctx, "*")
-	if err != nil {
-		return nil, fmt.Errorf("couldn't retrieve datastore list: %w", err)
-	}
-
-	return datastoreList, nil
-}
-
-// Precedence if not infraManagementUser:
-// * User from cluster
-// * User from Secret
-// Precedence if infraManagementUser:
-// * User from clusters infraManagementUser
-// * User from cluster
-// * User form clusters secret infraManagementUser
-// * User from clusters secret.
-func getUsernameAndPassword(cloud kubermaticv1.CloudSpec, secretKeySelector provider.SecretKeySelectorValueFunc, infraManagementUser bool) (username, password string, err error) {
-	if infraManagementUser {
-		username = cloud.VSphere.InfraManagementUser.Username
-		password = cloud.VSphere.InfraManagementUser.Password
-	}
-	if username == "" {
-		username = cloud.VSphere.Username
-	}
-	if password == "" {
-		password = cloud.VSphere.Password
-	}
-
-	if username != "" && password != "" {
-		return username, password, nil
-	}
-
-	if cloud.VSphere.CredentialsReference == nil {
-		return "", "", errors.New("cluster contains no password an and empty credentialsReference")
-	}
-
-	if username == "" && infraManagementUser {
-		username, err = secretKeySelector(cloud.VSphere.CredentialsReference, resources.VsphereInfraManagementUserUsername)
-		if err != nil {
-			return "", "", err
-		}
-	}
-	if username == "" {
-		username, err = secretKeySelector(cloud.VSphere.CredentialsReference, resources.VsphereUsername)
-		if err != nil {
-			return "", "", err
-		}
-	}
-
-	if password == "" && infraManagementUser {
-		password, err = secretKeySelector(cloud.VSphere.CredentialsReference, resources.VsphereInfraManagementUserPassword)
-		if err != nil {
-			return "", "", err
-		}
-	}
-
-	if password == "" {
-		password, err = secretKeySelector(cloud.VSphere.CredentialsReference, resources.VspherePassword)
-		if err != nil {
-			return "", "", err
-		}
-	}
-
-	if username == "" {
-		return "", "", errors.New("unable to get username")
-	}
-
-	if password == "" {
-		return "", "", errors.New("unable to get password")
-	}
-
-	return username, password, nil
-}
-
-func GetCredentialsForCluster(cloud kubermaticv1.CloudSpec, secretKeySelector provider.SecretKeySelectorValueFunc, dc *kubermaticv1.DatacenterSpecVSphere) (string, string, error) {
-	var username, password string
-	var err error
-
-	// InfraManagementUser from Datacenter
-	if dc != nil && dc.InfraManagementUser != nil {
-		if dc.InfraManagementUser.Username != "" && dc.InfraManagementUser.Password != "" {
-			return dc.InfraManagementUser.Username, dc.InfraManagementUser.Password, nil
-		}
-	}
-
-	// InfraManagementUser from Cluster
-	username, password, err = getUsernameAndPassword(cloud, secretKeySelector, true)
-	if err != nil {
-		return "", "", err
-	}
-
-	return username, password, nil
-}
-
-func ValidateCredentials(ctx context.Context, dc *kubermaticv1.DatacenterSpecVSphere, username, password string, caBundle *x509.CertPool) error {
-	session, err := newSession(ctx, dc, username, password, caBundle)
-	if err != nil {
-		return err
-	}
-	defer session.Logout(ctx)
-
-	_, err = session.Finder.DefaultFolder(ctx)
-
-	return err
 }
